Add a Compass enum
