+++ src/test/java/com/yesthisispatrick/games/minesweeper/TileTypeTest.java
@@ -24,24 +24,4 @@ public class TileTypeTest {
   public void testIncrementMax() {
     assertEquals(TILE_TYPE.EIGHT, TILE_TYPE.EIGHT.increment());
   }
-
-  @Test
-  public void testSouth() {
-    assertEquals(18, COMPASS.SOUTH.getPosition(8, 10));
-  }
-
-  @Test
-  public void testWest() {
-    assertEquals(8, COMPASS.WEST.getPosition(9, 10));
-  }
-
-  @Test
-  public void testSouthEast() {
-    assertEquals(19, COMPASS.SOUTHEAST.getPosition(8, 10));
-  }
-
-  @Test
-  public void testNorthWest() {
-    assertEquals(11, COMPASS.NORTHWEST.getPosition(22, 10));
-  }
 }
